BOOK-150
refactor(services): reduz complexidade e separa responsabilidades no bookService

- Extrai validações e construção de filtros para utilitários externos
- Centraliza tratamento de erros
- Organiza métodos CRUD e delega responsabilidades específicas para funções auxiliares
- Melhora legibilidade e manutenção do serviço de livros
